Run PrometheusVersionUpgrade tests separately
As our compatibility matrix grows, the tests assuring we can upgrade one Prometheus version to the next are taking up to 1~2h to complete. We aim to accelerate our e2e tests that are required to pass on Pull Requests by moving this test to a separate CI, that runs once a day.

Signed-off-by: Arthur Silva Sens <arthur.sens@coralogix.com>